LayoutTests/inspector/debugger/command-line-api-exception-expected.txt

CONSOLE MESSAGE: line 4: TypeError: undefined is not an object (evaluating '[].x.x')
CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: variableThatDoesNotExist
CONSOLE MESSAGE: line 16: SyntaxError: Unexpected token ')'
CONSOLE MESSAGE: line 22: IndexSizeError: DOM Exception 1: Index or size was negative, or greater than the allowed value.
CONSOLE MESSAGE: line 27: thrown string
CONSOLE MESSAGE: line 32: 123.456
CONSOLE MESSAGE: line 37: null
CONSOLE MESSAGE: line 42: [object Object]
CONSOLE MESSAGE: line 47: [object HTMLBodyElement]
CONSOLE MESSAGE: line 55: CATCHER: TypeError: undefined is not an object (evaluating '[].x.x')
CONSOLE MESSAGE: line 55: CATCHER: thrown string
CONSOLE MESSAGE: line 55: CATCHER: [object Object]
Checks that $exception is available and accurate in evaluations when paused on an exception.

BEFORE: $exception => undefined
PAUSE #1: $exception => TypeError: undefined is not an object (evaluating '[].x.x')
PAUSE #2: $exception => ReferenceError: Can't find variable: variableThatDoesNotExist
PAUSE #3: $exception => SyntaxError: Unexpected token ')'
PAUSE #4: $exception => Error: IndexSizeError: DOM Exception 1
PAUSE #5: $exception => thrown string
PAUSE #6: $exception => 123.456
PAUSE #7: $exception => null
PAUSE #8: $exception => Object
PAUSE #9: $exception => <body>
PAUSE #10: $exception => TypeError: undefined is not an object (evaluating '[].x.x')
STEPPED OUT TO CATCH BLOCK: $exception === e ? true
PAUSE #11: $exception => thrown string
STEPPED OUT TO CATCH BLOCK: $exception === e ? true
PAUSE #12: $exception => Object
STEPPED OUT TO CATCH BLOCK: $exception === e ? true
AFTER: $exception => undefined
